FE-73♻️  Tag관리 함수 훅으로 분리
